FilterFragment + MVVM

+++ app/src/main/java/com/example/task1/room/TableModel.kt
@@ -15,7 +15,7 @@ data class TableModel(
     val capital: String,
 
     @ColumnInfo
-    val area: Double,
+    val area: Float,
 
     @ColumnInfo
     val language: String,
